MortalityTables.jl

MortalityTables.Balducci

MortalityTables.Constant

MortalityTables.DeathDistribution

MortalityTables.MortalityTable

MortalityTables.SelectMortality

MortalityTables.UltimateMortality

MortalityTables.Uniform

MortalityTables.get_SOA_table

MortalityTables.get_SOA_table!

MortalityTables.p

MortalityTables.p

MortalityTables.p

MortalityTables.p

MortalityTables.p

MortalityTables.q

MortalityTables.q

MortalityTables.q_select

MortalityTables.q_ultimate

MortalityTables.tables

MortalityTables.Balducci — Type

Balducci()

A DeathDistribution type that assumes a decreasing force of mortality over the year.

MortalityTables.Constant — Type

Constant()

A DeathDistribution type that assumes a constant force of mortality over the year.

MortalityTables.DeathDistribution — Type

DeathDistribution

An abstract type used to form an assumption of how deaths occur throughout a year. See Balducci(), Uniform(), and Constant() for concrete assumption types.

MortalityTables.MortalityTable — Type

MortalityTable

A struct that holds a select (two-dimensional) and ultimate (vector) rates,
    along with MetaData associated with the table.

MortalityTables.SelectMortality — Type

Given an 2D array, will create a dictionary that is indexed by issue age and will return missing `if the age is not available.

MortalityTables.UltimateMortality — Type

Given an ultimate vector, will create a dictionary that is indexed by issue age and will return missing `if the age is not available.

MortalityTables.Uniform — Type

Uniform()

A DeathDistribution type that assumes an increasing force of mortality over the year.

MortalityTables.get_SOA_table! — Method

get_SOA_table!(dict,id)

Will lookup the given mortality table and add it to the given dict, with the name of the table acting as the added key in the dictionary.

This modifies the given dict (as is indicated by the conventional ! at the end of the function name).

!! Remember that not all tables have been tested to work.

MortalityTables.get_SOA_table — Method

get_SOA_table(id)

Given the id of a mort.SOA.org table, grab it and return it as a MortalityTable.

!! Remember that not all tables have been tested to work.

MortalityTables.p — Method

p(MortalityTable,issue_age,duration,time,DeathDistribution)

A function to calculate non-whole year survivorship, where you must make an assumption (DeathDistribution) about how the annual rate of mortaility applies through the rest of the year.

MortalityTables.p — Method

p(MortalityTable,issue_age,duration,time,DeathDistribution)

A function to calculate non-whole year survivorship, where you must make an assumption (DeathDistribution) about how the annual rate of mortaility applies through the rest of the year.

MortalityTables.p — Method

p(MortalityTable,issue_age,duration,time,DeathDistribution)

A function to calculate non-whole year survivorship, where you must make an assumption (DeathDistribution) about how the annual rate of mortaility applies through the rest of the year.

MortalityTables.p — Method

The probability that a life with given issue_age and currently in its nth durationdies survives to at least duration + time. If given select mortality, will be based on select rates.

Equivalant actuarial notation: $$_tp_{(x)+s}$$, or the probability that a life aged x + s who was select at age x survives to at least age x+s+t

MortalityTables.p — Method

the probability that a life aged issue_age + duration - 1 survives one additional timepoint

Equivalent actuarial notation: $$p_x$$ , or

MortalityTables.q — Method

q(MortalityTable,issue_age,duration,time,DeathDistribution)

A function to calculate non-whole year force of mortality, where you must make an assumption (DeathDistribution) about how the annual rate of mortaility applies through the rest of the year.

MortalityTables.q — Method

The probability that a life with given issue_age and currently in its nth durationdies by at least duration + time. If given select mortality, will be based on select rates.

Equivalent actuarial notation: $$p_{(x)+s}$$ or the probability that a life aged x + s who was select at age x dies by least age x+s+t

MortalityTables.q_select — Method

q_select(table::XTbMLTable, issueAge, duration)

Given a mortality table, an issue age, and a duration, returns the appropriate select or ultimate qx.

MortalityTables.q_ultimate — Method

q_ulitmate(table::XTbMLTable, age)

Given a mortality table and an age returns the appropriate ultimate qx.

MortalityTables.tables — Function

tables(dir=nothing)

Loads the XtbML (the SOA XML data format for mortality tables) stored in the
given path. If no path is specified, will load the packages in the
MortalityTables package directory. To see where your system keeps packages,
run `DEPOT_PATH` from a Julia REPL.