MortalityTables.jl

MortalityTables.Balducci

MortalityTables.Constant

MortalityTables.DeathDistribution

MortalityTables.Makeham

MortalityTables.MortalityTable

MortalityTables.TableMetaData

MortalityTables.Uniform

MortalityTables.Gompertz

MortalityTables.SelectMortality

MortalityTables.UltimateMortality

MortalityTables.cumulative_decrement

MortalityTables.get_SOA_table

MortalityTables.get_SOA_table!

MortalityTables.omega

MortalityTables.readXTbML

MortalityTables.survivorship

MortalityTables.tables

MortalityTables.μ

MortalityTables.Balducci — Type

Balducci()

A DeathDistribution type that assumes a decreasing force of mortality over the year.

MortalityTables.Constant — Type

Constant()

A DeathDistribution type that assumes a constant force of mortality over the year.

MortalityTables.DeathDistribution — Type

DeathDistribution

An abstract type used to form an assumption of how deaths occur throughout a year. See Balducci(), Uniform(), and Constant() for concrete assumption types.

MortalityTables.Makeham — Type

Makeham(a,b,c)

Construct a mortality model following Makeham's law:

$\mu_x = A + Bc^x$

MortalityTables.MortalityTable — Type

MortalityTable(ultimate)
MortalityTable(select, ultimate)
MortalityTable(select, ultimate; metadata::MetaData)

Constructs a container object which can hold either: - ultimate-only rates (an UltimateTable) - select and ultimate rates (a SelectUltimateTable)

Also pass a keyword argument metadata=MetaData(...) to store relevant information (source, notes, etc) about the table itself.

Examples

# first construct the underlying data
ult = UltimateMortality([x / 100 for x in 0:100]); # first ma
matrix = rand(10,50); # represents random mortality rates with a select period of 10 years
sel = SelectMortality(matrix,ult,start_age=0);

table = MortalityTable(sel,ult)

# can now get rates, indexed by attained age:

table.select[10] # the vector of rates for a risk issued select at age 10 

table.ultimate[99] # 0.99

MortalityTables.TableMetaData — Type

TableMetaData(kwargs...)

Has the following fields, which default to nothing if not specified with a keyword:

name - a name for the table

id - if a mort.SOA.org sourced table, will be the identifying table ID

provider - Where the rates came from

reference - Source for more info on table

content_type

description

comments

source_path

When you call a MortalityTable interactively, it will nicely print this summary infomration.

Example content from mort.SOA.org:

Table Identity: 1076

Provider Domain: actuary.org

Provider Name: American Academy of Actuaries

Table Reference: Tillinghast, “American Council of Life Insurers: Preferred Version of 2001 CSO Mortality Tables”, ...

Content Type: CSO/CET

Table Name: 2001 CSO Super Preferred Select and Ultimate - Male Nonsmoker, ANB

Table Description: 2001 Commissioners Standard Ordinary (CSO) Super Preferred Select and Ultimate Table – Male Nonsmoker. Basis: Age Nearest Birthday. Minimum Select Age: 0. Maximum Select Age: 99. Minimum Ultimate Age: 16. Maximum Ultimate Age: 120

Comments: Study Data: A preferred version of the 2001 Commissioners Standard Ordinary (CSO) table ...

And the source_path would be: https://mort.soa.org/ViewTable.aspx?&TableIdentity=1076

Example usage:

julia-repl> TableMetaData(name="My Table Name")
TableMetaData("My Table Name", nothing, nothing, nothing, nothing, nothing, nothing, nothing)

MortalityTables.Uniform — Type

Uniform()

A DeathDistribution type that assumes an increasing force of mortality over the year.

MortalityTables.Gompertz — Method

Gompertz(b,c)

Construct a mortality model following Gompertz' law of mortality:

$\mu_x = Bc^x$

This is a special case of Makeham's law $\mu_x = A + Bc^x$, where $A=0$.

Calling this will create a Makeham model that can be called with q or p.

MortalityTables.SelectMortality — Method

SelectMortality(select, ultimate; start_age=0)

Given a matrix rates, where the first row represents the select rates for a risk, will create a an OffsetArray that is indexed by issue age, containing a vector of rate indexed by attained age. The ultimate mortality vector is used for rates in the post-select period.

Give the optional keyword argument to start the indexing at an age other than zero.

Examples

julia> ult = UltimateMortality([x / 100 for x in 0:100]);

julia> matrix = rand(50,10); # represents random(!) mortality rates with a select period of 10 years

julia> sel = SelectMortality(matrix,ult,start_age=0);

julia> sel[0] # the mortality vector for a select life with issue age 0
 0.12858960119349439
 0.1172480189376135
 0.8237661916705163
 ⋮
 0.98
 0.99
 1.0

julia> sel[0][95] # the mortality rate for a life age 95, that was issued at age 0
0.95

MortalityTables.UltimateMortality — Method

UltimateMortality(vector; start_age=0)

Given a vector of rates, returns an OffsetArray that is indexed by attained age.

Give the optional keyword argument to start the indexing at an age other than zero.

Examples

julia> m = UltimateMortality([0.1,0.3,0.6,1]);

julia> m[0]
0.1

julia> m = UltimateMortality([0.1,0.3,0.6,1], start_age = 18);

julia> m[18]
0.1

MortalityTables.cumulative_decrement — Method

cumulative_decrement(mortality_vector,to_age)
cumulative_decrement(mortality_vector,from_age,to_age)

Returns the cumulative decrement through attained age to_age. The start of the calculation is either the start of the vector, or attainedage `fromage.fromageandtoage` need to be Integers. Add a DeathDistribution as the last argument to handle floating point and non-whole ages:

cumulative_decrement(mortality_vector,to_age,::DeathDistribution)
cumulative_decrement(mortality_vector,from_age,to_age,::DeathDistribution)

Examples

julia> qs = UltimateMortality([0.1,0.3,0.6,1]);
    
julia> cumulative_decrement(qs,0)
0.0
julia> cumulative_decrement(qs,1)
0.1

julia> cumulative_decrement(qs,1,1)
0.0
julia> cumulative_decrement(qs,1,2)
0.3

julia> cumulative_decrement(qs,0.5,Uniform())
0.05

MortalityTables.get_SOA_table! — Method

get_SOA_table!(dict,id)

Will lookup the given mortality table and add it to the given dict, with the name of the table acting as the added key in the dictionary.

This modifies the given dict (as is indicated by the conventional ! at the end of the function name).

!! Remember that not all tables have been tested to work.

MortalityTables.get_SOA_table — Method

get_SOA_table(id)

Given the id of a mort.SOA.org table, grab it and return it as a MortalityTable.

!! Remember that not all tables have been tested to work.

MortalityTables.omega — Method

omega(x)
ω(x)

Returns the last index of the given vector. For mortality vectors this means the last attained age for which a rate is defined.

ω is aliased to omega, but unexported. To use, do using MortalityTables: ω when importing or call MortalityTables.ω()

Examples

julia> qs = UltimateMortality([0.1,0.3,0.6,1]);
julia> omega(qs)
3

julia> qs = UltimateMortality([0.1,0.3,0.6,1],start_age=10);
julia> omega(qs)
13

MortalityTables.survivorship — Method

survivorship(mortality_vector,to_age)
survivorship(mortality_vector,from_age,to_age)

Returns the survivorship through attained age to_age. The start of the calculation is either the start of the vector, or attainedage `fromage.fromageandtoage` need to be Integers. Add a DeathDistribution as the last argument to handle floating point and non-whole ages:

survivorship(mortality_vector,to_age,::DeathDistribution)
survivorship(mortality_vector,from_age,to_age,::DeathDistribution)

If given a negative to_age, it will return 1.0. Aside from simplifying the code, this makes sense as for something to exist in order to decrement in the first place, it must have existed and surived to the point of being able to be decremented.

Examples

julia> qs = UltimateMortality([0.1,0.3,0.6,1]);
    
julia> survivorship(qs,0)
1.0
julia> survivorship(qs,1)
0.9

julia> survivorship(qs,1,1)
1.0
julia> survivorship(qs,1,2)
0.7

julia> survivorship(qs,0.5,Uniform())
0.95

MortalityTables.readXTbML — Method

readXTbML(path)

Loads the [XtbML](https://mort.soa.org/About.aspx) (the SOA XML data format for mortality tables) stored at the given path and returns a `MortalityTable`.

MortalityTables.tables — Function

tables(dir=nothing)

Loads the XtbML (the SOA XML data format for mortality tables) stored in the given path. If no path is specified, will load the packages in the MortalityTables package directory. To see where your system keeps packages, run DEPOT_PATH from a Julia REPL.

MortalityTables.μ — Method

μ(m::ParametricMortality,age)

$\mu_x$: Return the force of mortality at the given age.